Make linked workspace error handling decorator reusable.
The error handling done for the LinkedWorkspacesService endpoints
is generic for endpoints proxying to other services. We wrap it in
an external function allowing to create decorators specific for
a given service.
